backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.2
groq==0.9.0
resend==2.2.0
slowapi==0.1.9
python-magic==0.4.27
pydantic[email]==2.7.4
python-dotenv==1.0.1